string array read/write + unit tests for string/string array read/write

+++ include/virg/nexus/nx_string.h
@@ -26,8 +26,8 @@ void nx_strredup(char ** dest, const char *src);
 char *nx_fstr(const char* format, ...) __attribute__((format (printf, 1, 2)));
 char *nx_vfstr(const char* format, va_list args);
 
-size_t nx_strwrite(const char *s, FILE *stream);
-size_t nx_strread(char **s, size_t size, FILE *stream);
+int nx_strwrite(const char *s, FILE *stream);
+int nx_strread(char **s, size_t size, FILE *stream);
 char *nx_strread0(FILE *stream);
 
 __NX_END_DECL
